Move enum utilities to @odh-dashboard/foundation
