+++ frontend/src/components/OdhDocListItem.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Button, Tooltip, Label } from '@patternfly/react-core';
 import { ExternalLinkAltIcon } from '@patternfly/react-icons';
+import { asEnumMember } from '@odh-dashboard/foundation';
 import { OdhDocument, OdhDocumentType } from '#~/types';
 import { getQuickStartLabel, launchQuickStart } from '#~/utilities/quickStartUtils';
 import { DOC_TYPE_TOOLTIPS } from '#~/utilities/const';
-import { getLabelColorForDocType, asEnumMember, getDuration } from '#~/utilities/utils';
+import { getLabelColorForDocType, getDuration } from '#~/utilities/utils';
 import { useQuickStartCardSelected } from './useQuickStartCardSelected';
 import FavoriteButton from './FavoriteButton';
