+++ frontend/src/concepts/pipelines/content/createRun/RunPage.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
 import { PageSection } from '@patternfly/react-core';
+import { asEnumMember } from '@odh-dashboard/foundation';
 import useIsMlflowPipelinesAvailable from '#~/concepts/mlflow/hooks/useIsMlflowPipelinesAvailable';
 import { ExperimentKF, PipelineRecurringRunKF, PipelineRunKF } from '#~/concepts/pipelines/kfTypes';
 import GenericSidebar from '#~/components/GenericSidebar';
@@ -22,7 +23,6 @@ import {
 import { ValueOf } from '#~/typeHelpers';
 import { useGetSearchParamValues } from '#~/utilities/useGetSearchParamValues';
 import { PipelineRunSearchParam } from '#~/concepts/pipelines/content/types';
-import { asEnumMember } from '#~/utilities/utils';
 import useDefaultExperiment from '#~/pages/pipelines/global/experiments/useDefaultExperiment';
 
 type RunPageProps = {
