+++ frontend/src/pages/connectionTypes/manage/ConnectionTypeDataFieldModal.tsx
@@ -12,6 +12,7 @@ import {
   Alert,
 } from '@patternfly/react-core';
 import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons';
+import { isEnumMember } from '@odh-dashboard/foundation';
 import { DashboardPopupIconButton } from '@odh-dashboard/ui-core';
 import ContentModal from '#~/components/modals/ContentModal';
 import {
@@ -27,7 +28,6 @@ import {
   isModelServingEnvVar,
   isValidEnvVar,
 } from '#~/concepts/connectionTypes/utils';
-import { isEnumMember } from '#~/utilities/utils';
 import DataFieldPropertiesForm from '#~/pages/connectionTypes/manage/DataFieldPropertiesForm';
 import { prepareFieldForSave } from '#~/pages/connectionTypes/manage/manageFieldUtils';
 import useGenericObjectState from '#~/utilities/useGenericObjectState';
