+++ frontend/src/pages/modelServing/screens/metrics/bias/BiasConfigurationPage/BiasConfigurationModal/MetricTypeField.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { FormGroup } from '@patternfly/react-core';
+import { asEnumMember, enumIterator } from '@odh-dashboard/foundation';
 import {
   METRIC_TYPE_DESCRIPTION,
   METRIC_TYPE_DISPLAY_NAME,
 } from '#~/pages/modelServing/screens/metrics/const';
 import { BiasMetricType } from '#~/api';
 import { isMetricType } from '#~/pages/modelServing/screens/metrics/utils';
-import { asEnumMember, enumIterator } from '#~/utilities/utils';
 import SimpleSelect, { SimpleSelectOption } from '#~/components/SimpleSelect';
 
 type MetricTypeFieldProps = {
