+++ frontend/src/pages/modelServing/screens/metrics/utils.tsx
@@ -3,6 +3,7 @@ import { ReactElement } from 'react';
 import * as _ from 'lodash-es';
 import { BreadcrumbItem } from '@patternfly/react-core';
 import { Link } from 'react-router-dom';
+import { isEnumMember } from '@odh-dashboard/foundation';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import { TimeframeTitle } from '#~/concepts/metrics/types';
 import { BreadcrumbItemType, PrometheusQueryRangeResultValue } from '#~/types';
@@ -13,7 +14,6 @@ import {
   BIAS_THRESHOLD_COLOR,
 } from '#~/pages/modelServing/screens/metrics/const';
 import { PROMETHEUS_REQUEST_RESOLUTION } from '#~/concepts/metrics/const';
-import { isEnumMember } from '#~/utilities/utils';
 import {
   BiasSelectOption,
   DomainCalculator,
