+++ frontend/src/pages/notebookController/screens/admin/NotebookControllerTabs.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { Tab, Tabs, TabTitleText } from '@patternfly/react-core';
+import { asEnumMember } from '@odh-dashboard/foundation';
 import { NotebookControllerTabTypes } from '#~/pages/notebookController/const';
 import NotebookServerRoutes from '#~/pages/notebookController/screens/server/NotebookServerRoutes';
 import { NotebookControllerContext } from '#~/pages/notebookController/NotebookControllerContext';
-import { asEnumMember } from '#~/utilities/utils';
 import NotebookAdmin from './NotebookAdmin';
 
 const NotebookControllerTabs: React.FC = () => {
