+++ frontend/src/pages/pipelines/global/runs/GlobalPipelineRunsTabs.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { PageSection, Tab, Tabs, TabTitleText } from '@patternfly/react-core';
-import { asEnumMember } from '#~/utilities/utils';
+import { asEnumMember } from '@odh-dashboard/foundation';
 import ScheduledRuns from '#~/pages/pipelines/global/runs/ScheduledRuns';
 import { ActiveRuns } from '#~/pages/pipelines/global/runs/ActiveRuns';
 import { ArchivedRuns } from '#~/pages/pipelines/global/runs/ArchivedRuns';
