+++ frontend/src/pages/projects/screens/spawner/environmentVariables/EnvSecret.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
+import { asEnumMember } from '@odh-dashboard/foundation';
 import { EnvironmentVariableType, EnvVariableData, SecretCategory } from '#~/pages/projects/types';
-import { asEnumMember } from '#~/utilities/utils';
 import EnvDataTypeField from './EnvDataTypeField';
 import GenericKeyValuePairField from './GenericKeyValuePairField';
 import { EMPTY_KEY_VALUE_PAIR } from './const';
