+++ packages/foundation/jest.config.ts
@@ -1,0 +1,3 @@
+import baseConfig from '@odh-dashboard/jest-config';
+
+export default baseConfig;
