+++ packages/foundation/src/index.ts
@@ -1,1 +1,2 @@
+export { asEnumMember, enumIterator, isEnumMember, isInEnum } from './enumUtils';
 export { genRandomChars } from './utils';
