+++ packages/model-serving/package.json
@@ -32,10 +32,10 @@
     "./utils": "./src/concepts/utils.ts"
   },
   "dependencies": {
+    "@odh-dashboard/foundation": "*",
     "@odh-dashboard/internal": "*",
     "@odh-dashboard/k8s-core": "*",
     "@odh-dashboard/model-registry": "*",
-
     "@odh-dashboard/plugin-core": "*",
     "@odh-dashboard/ui-core": "*",
     "lodash-es": "^4.18.1",
